added languages, editors, OS's and machines pie chart with click interaction, add colorHash function and a list of programming languages colors for the languages pie chart, fixed last 7 days line chart re-rendering on configuration change (like theme switch)

+++ app/src/main/java/com/stefdp/hackatime/network/hackatimeapi/models/Language.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Language(
     override val name: String,
-    @SerializedName("total_seconds") override val totalSeconds: Int,
+    @SerializedName("total_seconds") override val totalSeconds: Double,
     override val text: String,
     override val hours: Int,
     override val minutes: Int,
