+++ app/src/main/java/com/stefdp/hackatime/network/hackatimeapi/models/Project.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Project(
     override val name: String,
-    @SerializedName("total_seconds") override val totalSeconds: Int,
+    @SerializedName("total_seconds") override val totalSeconds: Double,
     override val text: String,
     override val hours: Int,
     override val minutes: Int,
